# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NFT generator",
    "description": "This is a sample server for a NFT generator.",
    "termsOfService": "http://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server",
      "variables": {
        "port": {
          "default": "8000"
        },
        "basePath": {
          "default": "/"
        }
      }
    }
  ],
  "components":{
     "securitySchemes":{
        "authorization":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}
     }
  },
    "security":[{ "authorization":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}],
  "paths": {
...
    "/createCollection": {
      "post": {
        "summary": "",
        "operationId": "createCollection",
        "parameters": [], 
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "collection": {
                    "description": "collection name",
                    "type": "number"
                  },
                    "height": {
                      "description": "height",
                      "type": "number"
                    },
                    "width":{
                      "description": "width",
                      "type": "number"
                    }
                    
                },
                "required": ["collection","height","width"] 
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CREATED"
          },
          "405": {
            "description": "Method Not Allowed"
          }
        }
      }
    },
...
  }
}